Removed collection cascading solving the main issue from past few weeks
I do not know what effect this will have in the future (cascading sounds
like something we would want) but doing this solved the issue with
NonUniqueObjectException I've been battling with for the last 2-3 weeks
so I'm taking whatever problems may arise in the future in exchange for
this solution.

I also tried enabling lazy loading at one point but that caused a whole
another set of problems with LazyInitializationException. Note for
future: lazy loading is really useful but apparently cannot be
implemented in it's default state in this project. Consider researching
the use of manual joins.

Another potential solution I attempted was fiddling with the session &
transaction management but that didn't solve anything as that was not
the problem. I've left my changes to the session/transaction management
I did while attempting to solve this problem as the changes did not
actually seem to change a thing.